fix a bug that mac not support f64

+++ python/taichi/ui/window.py
@@ -183,7 +183,7 @@ class Window:
         """
         tmp_depth = get_depth_ndarray(self.window)
         self.window.copy_depth_buffer_to_ndarray(tmp_depth.arr)
-        depth_numpy_arr = numpy.zeros(self.get_window_shape())
+        depth_numpy_arr = numpy.zeros(self.get_window_shape(), dtype=numpy.float32)
         arr_vulkan_layout_to_arr_normal_layout(tmp_depth, depth_numpy_arr)
         return depth_numpy_arr
